requirements.txt
chardet==5.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
psutil==6.1.1
PyYAML==6.0.2
reportlab==4.3.0
pyyaml
jinja2
reportlab
psutil
wmi
pywin32